+++ tests/testthat/test-1-data-structures.R
@@ -241,17 +241,17 @@ context("faunistické čtverce")
 
   expect_error(KFME_grid("bflm")) # neznámé rozlišení - očekávám high(default) / low
 
-  telc <- geocode("Telč") %>%
-    filter(typ == "Obec") %>%  # bod Telč
-    st_set_agr("constant")
+  telc <- geocode("Telč") %>% # známý bod Telč
+    filter(typ == "Obec")
 
-  hrcava <- geocode("Hrčava") %>%
-    filter(typ == "Obec") %>%  # bod Hrčava
-    st_set_agr("constant")
 
-  cernousy <- geocode("Černousy") %>%
-    filter(typ == "Obec") %>%  # bod Hrčava
-    st_set_agr("constant")
+  hrcava <- geocode("Hrčava") %>% # známý bod Hrčava
+    filter(typ == "Obec")
+
+
+  cernousy <- geocode("Černousy") %>% # známý bod Černousy
+    filter(typ == "Obec")
+
 
   expect_equal(sf::st_intersection(KFME_grid("low"), telc)$ctverec, 6858) # bod Telč je ve velkém čtverci 6858
   expect_equal(sf::st_intersection(KFME_grid("high"), telc)$ctverec, "6858b") # bod Telč je v malém čtverci 6858b