+++ NEWS.md
@@ -1,5 +1,7 @@
 ## version 1.4.2
-  -  KFME grid cells (faunistické čtverce) object added
+  - KFME grid cells (faunistické čtverce) object added
+  
+  - minor optimization of geocode() and tests 
 
 ## version 1.4.1